rename librepr::detail to librepr::util
