+++ benchmark/runtime/util/string_buffer.cpp
@@ -5,7 +5,7 @@
 
 #include <librepr/util/string/buffer.h>
 
-namespace librepr::detail::benchmark {
+namespace librepr::util::benchmark {
 void string_append(::benchmark::State& state) {
   for (auto _ : state) {
     std::string buffer{};
@@ -36,7 +36,7 @@ BENCHMARK(stringstream_append);
 
 void stringbuffer_append(::benchmark::State& state) {
   for (auto _ : state) {
-    librepr::detail::StringBuffer buffer{};
+    librepr::util::StringBuffer buffer{};
     for (unsigned i = 0; i < 100; ++i){
       buffer << "fooabc1234567890 asdf <>()";
     }
